requirements.txt
argcomplete==1.1.1
boto==2.39.0
boto3==1.3.0
botocore==1.4.10
cement==2.8.2
Cerberus==0.9.2
colorama>=0.3.3
docker-py==1.8.0
docopt==0.6.2
docutils==0.12
Jinja2==2.8
jmespath==0.9.0
MarkupSafe==0.23
ordereddict==1.1
pkginfo==1.2.1
prompt-toolkit==0.60
pyaml==15.8.2
Pygments==2.1.3
pyparsing==2.1.1
pyroma==2.0.2
python-dateutil==2.5.2
PyYAML==3.11
requests==2.9.1
schema==0.4.0
six==1.10.0
termcolor==1.1.0
wcwidth==0.1.6
websocket-client==0.35.0
